add reusable `DisableAnimationsInTestEnvironment` middleware
this middleware was already generated into the `internal_test_app` test
application, but was otherwise fenced off from users. this moves it to a place
where inheriting applications can use it.

including this middleware with `config.middleware.use
DisableAnimationsInTestEnvironment` should improve Capybara's test reliability
by stripping out transition animations.
