+++ .dassie/config/environments/test.rb
@@ -1,3 +1,5 @@
+require 'hyrax/specs/disable_animations_in_test_environment'
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -43,4 +45,7 @@ Rails.application.configure do
 
   # Raises error for missing translations
   config.action_view.raise_on_missing_translations = true
+
+  # disable animations with shared hyrax code
+  config.middleware.use DisableAnimationsInTestEnvironment
 end
